requirements.txt
# Django Multi-Service Platform Requirements (MongoDB Version)

# Core Django
Django==4.2.15
djangorestframework==3.14.0
django-cors-headers==4.3.1

# MongoDB Integration
mongoengine==0.27.0
django-mongoengine==0.5.4
pymongo==4.5.0
djongo==1.3.6  # Alternative MongoDB connector

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.4
mongoengine-jwt==0.1.1  # JWT for MongoEngine

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0

# API Documentation
drf-yasg==1.21.7

# Image & File Handling
Pillow==10.0.0
django-storages==1.14.2
boto3==1.28.62  # For AWS S3 storage
gridfs==0.0.1  # MongoDB GridFS for file storage

# Payment Integration
razorpay==1.3.0
stripe==6.6.0

# Real-time Features
channels==4.0.0
channels-redis==4.1.0
redis==4.6.0

# Notifications
pyfcm==1.5.4  # Firebase Cloud Messaging
twilio==8.9.1  # SMS notifications

# Location Services
geopy==2.4.0
geoip2==4.7.0

# Task Queue
celery==5.3.2
django-celery-beat==2.5.0

# Monitoring & Logging
sentry-sdk==1.32.0
django-extensions==3.2.3

# Development Tools
django-debug-toolbar==4.2.0
factory-boy==3.3.0  # For testing
pytest-django==4.5.2

# Utilities
requests==2.31.0
python-dateutil==2.8.2
phonenumbers==8.13.22
django-phonenumber-field==7.1.0

# Export & Reporting
openpyxl==3.1.2
reportlab==4.0.4

# Search (MongoDB compatible)
mongoengine-elasticsearch==0.1.0
elasticsearch==8.9.0

# Data Validation
marshmallow==3.20.1
marshmallow-mongoengine==0.30.0

# MongoDB Admin Tools
mongo-connector==3.1.1
mongodbtools==0.1.0

# Additional MongoDB utilities
bson==0.5.10
motor==3.3.1  # Async MongoDB driver